changes for abstracting pool away from endpoints
